+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+- ingest extracted organizational units on ldap search
 - ORCID search endpoint in mex-backend
   - endpoint returning orcid persons (tranformed to ExtractedPersons)
   - full-text search query parameter `q` and pagination parameters `offset` and `limit`